refactor(lowering): split expressions.rs into expr/ sub-module
