+++ crates/js-lowering/src/lower/expr/mod.rs
@@ -1,0 +1,7 @@
+mod assignment;
+mod binary;
+mod call;
+mod collection;
+mod function;
+mod literal;
+mod member;
